requirements.txt
Flask
flask-cors
python-dotenv
Werkzeug
openai
line-bot-sdk
gunicorn